Add new index.js to start server
We are moving closer to the example the Hapi docs provide for setting up and running the Hapi server.

https://hapi.dev/tutorials/testing/?lang=en_US

Specifically, the version that supports testing (because duh, of course we're testing) not the one in their 'Getting started' section.

Rather than try and be clever about having the code determine if it was just required or actually run, it expects the caller to make that determination by the method it calls.

First step is to add an `index.js` file to the root of the project and ensure it is our main file going forward.