Move plugin registration into its own method
This is just to make things clearer, and again, to bring the `init()` and `start()` methods inline with the Hapi example.

Note - The Hapi example does not include a call to `init()`. We tried not doing that but though the server ran, nothing was registered, not even the routes. So, we think this is an oversight in their example.