Fix issue with global server instance
We set `server` as a const in `server.js` as per the Hapi example. TBH, we're not sure how their example would work in the real world because we found a massive chunk of our tests were now failing.

The first use of `init()` would work. But any subsequent call would fail with an error telling us `HapiNowAuthPlugin` is already registered.

A review of the file spotted the problem. With `const server` being declared in the root of the module, the first call to `init()` would set the instance variable. All subsequent calls would be trying to interact with that same instance, which will already have had all the plugins registered and have been initialised.

So, we solved the problem by moving the server instance into the scope of `init()`. Now each call will get its own instance avoiding any conflicts in the tests.