Remove hpal support and update server.js (#467)
Back when we initialised the project we had the intention of using [hpal-debug](https://github.com/hapipal/hpal-debug) to support our work. The idea we really liked was being able to interact with our Objection models in a REPL. For those of us with Ruby & Rails backgrounds, the idea brought us much joy.

Fast forward a year and we've never actually taken the time to hook up Objection properly plus no one on the team can recall ever using the Hpal debugging tools.

So, to help keep the project lean and mean we're dropping it from the project, which includes updating `server.js to follow the pattern we see in other projects and the Hapi docs. Specifically, the example the Hapi docs provide for setting up and running the Hapi server.

https://hapi.dev/tutorials/testing/?lang=en_US

This version also supports testing (because duh, of course we're testing) as opposed to the one in their 'Getting started' section. Rather than try and be clever about having the code determine if it was just required or actually run, it expects the caller to make that determination by the method it calls. In tests you'd call `init()`, and to actually run the server you add an `index.js` and call `start()` from it.

** Notes

We move the plugin registration into its own method. This is just to make things clearer, and again, to bring the `init()` and `start()` methods inline with the Hapi example.

Also, the Hapi example does not include a call to `init()`. We tried not doing that but though the server ran, nothing was registered, not even the routes. So, we think this is an oversight in their example.

We also moved server.js into the app folder. With index.js now at the root it makes sense to us to move server.js into the `app/` folder with the rest of the application code.

Finally, another deviation from the Hapi example. Initially, we set `server` as a const in `server.js` as per the Hapi example. TBH, we're not sure how their example would work in the real world because we found a massive chunk of our tests were now failing. The first use of `init()` would work. But any subsequent call would fail with an error telling us `HapiNowAuthPlugin` is already registered. A review of the file spotted the problem. With `const server` being declared in the root of the module, the first call to `init()` would set the instance variable. All subsequent calls would be trying to interact with that same instance, which will already have had all the plugins registered and have been initialised. So, we solved the problem by moving the server instance into the scope of `init()`. Now each call will get its own instance avoiding any conflicts in the tests.