Delete all index.js files
This is being committed first but was actually towards the end of the process of getting the app back up and running using ES6 imports.

The key blocker was that when using ES6 imports we got a sort of circular dependency when it came to using `extends` with classes.

We'd grab the base class via `../index.js` but then it would complain the thing we were extending was not a class. Googling threw up little help. But we did identify that if we required the base class directly the problem as long as we did not import anything else using `../index.js` it worked. The problem was this left a good many files needing to be updated to import things directly, and others not needing to.

So, we took the bold decision that rather than half & half things, or have exceptions all other the place we would _always_ be explicit in our imports.

This why we are being consistent throughout and it means we can delete a good chunk of code. Also, it turns out when you do it this way VSCode recognises all the files and in most cases will auto-complete the import statement for you.

So, it is a complete shift in direction (we were intending to use `index.js` in sub-folders). But actually for new starters it hopefully makes things more verbose and therefore a little clearer to follow.

N.B. `index.js` in the `db/` folder is the exception. It is actually doing a bunch of logic and work. We probably should somehow move this to its own file but right now its is a bit too mcuh work to figure out how best to do that.