Update CognitoJwtToPemService
We are explicitly calling out the update to this class to use ES6 imports because we have had to make substantial changes to accomadate that you can't use ES6 `import` to read in a JSON file.

Once we realised this we knew we needed to switch to using `fs` directly. But it needed an absolute path to the file being read which lead to our next problem; meta variables like `__dirname` are not available in ES6 modules.

So, we also needed to add a means to identify the current directory and use `path` to ensure the path was correctly built before passing it to `fs`.

Ref: https://techsparx.com/nodejs/esnext/dirname-es-modules.html