App is now running
With the code in this state the app now starts and authentication appears to work. Base queries, for example, list regimes work as well. However, anything more complex fails.

This appears to be a limitation of Objection being used in a project where `"type": "module"` has been set in the `package.json`. We'll be attempting to fix it in the next commits.