Switch to using explicit model declarations
> Require loops (circular dependencies, circular requires) are a very common problem when defining relations. Whenever a module A imports module B that immediately (synchronously) imports module A, you create a require loop that node.js or objection cannot solve automatically.
>
> https://vincit.github.io/objection.js/guide/relations.html#require-loops

If you follow the link you'll see 3 suggested solutions to this issue. The 3rd is to use 'ModelPaths' https://vincit.github.io/objection.js/api/model/static-properties.html#static-modelpaths

This requires having a `BaseModel` and setting a getter property called `modelPaths` which defines where the models live in relation to the `BaseModel`. Typically they are in the same folder hence you get

```
class BaseModel extends Model {
  static get modelPaths() {
    return [__dirname];
  }
}
```

If you have read the previous commit messages you'll know that switching to ES6 modules means `__dirname` is no longer available to us. However, the critical problem is that the code Objection uses to resolve the models depends on `require()`.

```
require(`${path.resolve(modelPath)}`)
```

(See https://github.com/Vincit/objection.js/issues/2023#issuecomment-865803273). When left in you get a message from Node which TL;DR is please remove the `"type": "module"` in `package.json`.

ES6 modules also come to the rescue in this situation. If we instead use an altered version of solution 1 to the require loop problem we can both explicitly tell Objection what to use and resolve the critical problem.

Because ES6 modules don't have the circular dependency issue they don't fall foul of the require loop problem. So, we can declare the imports at the top of the file as we normally would knowing they'll just be called once and then grabbed from cache on subsequent calls.

You could also argue it removes some of the 'mystery' to how the models are linked and related queries work.