Fix JoiDate extension not working
When we made the change to ES6 import we admittedly made a best guess as to the way to extend Joi with the joi-date extension. There is simply nothing in the Joi docs on how to do it using the ES6 syntax.

Thankfully we tracked down this issue https://github.com/sideway/joi-date/issues/39#issue-864015391 which included how they had used `@joi/date` with Joi to extend it.

So, this updates our code to match and resolve the error we were getting (essentially Joi was not recognising our date formats).