Fix - Rebilling for same customer causes error
CMEA-64 https://eaflood.atlassian.net/browse/CMEA-64

During testing we encountered an issue we had not anticipated. If you attempt to rebill multiple invoices that happen to be for the same customer and financial year you get a DB conflict error.

```json
{
 "statusCode": 409,
 "error": "Conflict",
 "message": "UniqueViolationError - Key (bill_run_id, customer_reference, financial_year, rebilled_type)= (5e969d83-8f32-483f-b90b-601bdece7d37, T1234567890A, 2020, C) already exists."
}
```

The problem is the constraint we use on the `invoices` table. This is actually at the core of the logic we use when transactions are added, to determine when a new invoice record needs creating or an existing one needs adding too. When adding transactions we always attempt to add a new invoice and rely on PostgreSQL's `ON CONFLICT` functionality to tell us actually we just need to update the details of an existing invoice.

In [Create rebilling initialise service](https://github.com/DEFRA/sroc-charging-module-api/pull/395) we tweaked the constraint to take account of invoice type. This is what allows us to have 3 invoices for the same customer and financial year; as long as the types are all different. But what if a customer, who was invoiced on 2 separate bill runs, needs both rebilled in the next bill run? This could happen but our constraint is not written to handle it.

This fix updates the constraint to allow us to handle the scenario, whilst ensuring we don't affect existing behaviour.

Steps to reproduce

- Use `T1234567890A` and financial year `2020` for all transactions
- Create and complete bill run having added some transactions
  - Grab the Invoice ID for rebilling later
- Create and complete another bill run having added some more transactions
  - Grab the Invoice ID for rebilling later
- Create a new bill run
  - request to rebill the first invoice ID grabbed
  - request to rebill the second invoice ID grabbed (this request will error)