Interim commit - working solution
By adding the `rebilled_invoice_id` to the invoice constraint we get something that will allow multiple rebilling invoices for the same customer and financial year, whilst also throwing an `ON CONFLICT` error so our add transaction process behaves as expected (creating an invoice when no conflict, updating the existing one when there is a conflict).

The one issue we have found is that `rebilled_invoice_id` must be populated with something when transactions are added to the original invoice else `ON CONFLICT` doesn't seem to fire and you get an invoice per transaction. To confirm this solution would work we've gone with the current bill run ID. But that's be confusing from a reporting point of view so we need to think about this some more.