Record attempt at using DB default
We did attempt to set the `rebilledInvoiceId` field in the DB by using a default value for the column of '99999999-9999-9999-9999-999999999999'.

This worked fine for new invoices. The problem is the rebill functionality leans into the same UPSERT mechanism in order to update the tally values in the invoice. For this to work the `InvoiceModel._baseOnInsertObject()` now needs a `rebilledInvoiceId` to be passed through else the constraint won't fire (which triggers the update rather than insert).

When set at the database level the default is applied as part of the insert. To get our constraint to work correctly we always need to specify a `rebilledInvoiceId`. Which means if we always have to specify it, why bother with a DB default!?