Correct invoice rebilling initialise test
Looking to add a test here for our scenario and realised the existing test did not reflect what would actually happen.

The invoice being rebilled should be on a different bill run (the rebill bill run validator would ensure this). In the test, the same bill run was linked to the original and the rebill, and only because the validation is handled outside the service was this working.

This updates the test so the original invoice being rebilled is linked to a different bill run.