Fix - Rebilling for same customer causes error (#475)
CMEA-64 https://eaflood.atlassian.net/browse/CMEA-64

During testing we encountered an issue we had not anticipated. If you attempt to rebill multiple invoices that happen to be for the same customer and financial year you get a DB conflict error.

```json
{
 "statusCode": 409,
 "error": "Conflict",
 "message": "UniqueViolationError - Key (bill_run_id, customer_reference, financial_year, rebilled_type)= (5e969d83-8f32-483f-b90b-601bdece7d37, T1234567890A, 2020, C) already exists."
}
```

The problem is the constraint we use on the `invoices` table. This is actually at the core of the logic we use when transactions are added, to determine when a new invoice record needs creating or an existing one needs adding too. When adding transactions we always attempt to add a new invoice and rely on PostgreSQL's `ON CONFLICT` functionality to tell us actually we just need to update the details of an existing invoice.

In [Create rebilling initialise service](https://github.com/DEFRA/sroc-charging-module-api/pull/395) we tweaked the constraint to take account of invoice type. This is what allows us to have 3 invoices for the same customer and financial year; as long as the types are all different. But what if a customer, who was invoiced on 2 separate bill runs, needs both rebilled in the next bill run? This could happen but our constraint is not written to handle it.

This fix updates the constraint to allow us to handle the scenario, whilst ensuring we don't affect existing behaviour. It also includes a data migration (a first) to ensure any existing data is updated to match how new invoices will be set when created.

Steps to reproduce

- Use `T1234567890A` and financial year `2020` for all transactions
- Create and complete bill run having added some transactions
  - Grab the Invoice ID for rebilling later
- Create and complete another bill run having added some more transactions
  - Grab the Invoice ID for rebilling later
- Create a new bill run
  - request to rebill the first invoice ID grabbed
  - request to rebill the second invoice ID grabbed (this request will error)

* Notes

By adding the `rebilled_invoice_id` to the invoice constraint we get something that will allow multiple rebilling invoices for the same customer and financial year, whilst also throwing an `ON CONFLICT` error so our add transaction process behaves as expected (creating an invoice when no conflict, updating the existing one when there is a conflict).

The one issue we have found is that `rebilled_invoice_id` must be populated with something when transactions are added to the original invoice else `ON CONFLICT` doesn't seem to fire and you get an invoice per transaction. We found a valid UUID format that hopefully is recognisable as something set rather than generated so that is what we are setting the field to.

* Record attempt at using DB default

We did attempt to set the `rebilledInvoiceId` field in the DB by using a default value for the column of '99999999-9999-9999-9999-999999999999'.

This worked fine for new invoices. The problem is the rebill functionality leans into the same UPSERT mechanism in order to update the tally values in the invoice. For this to work the `InvoiceModel._baseOnInsertObject()` now needs a `rebilledInvoiceId` to be passed through else the constraint won't fire (which triggers the update rather than insert).

When set at the database level the default is applied as part of the insert. To get our constraint to work correctly we always need to specify a `rebilledInvoiceId`. Which means if we always have to specify it, why bother with a DB default!?

* Notes correcting the invoice rebilling initialise test

When looking to add a test here for our scenario we realised the existing test did not reflect what would actually happen.

The invoice being rebilled should be on a different bill run (the rebill bill run validator would ensure this). In the test, the same bill run was linked to the original and the rebill, and only because the validation is handled outside the service was this working.

So, we updated the test so the original invoice being rebilled is linked to a different bill run.