Create table body presenter
Similar to how we generate transaction and customer files, we will be piping the contents of a db table row by row through a presenter which formats the row into sequentially-numbered columns, and then through a presenter which formats the row as CSV, and finally piped into a file.

We start developing this process by creating the table body presenter. Note that we may end up reusing this for the table head; if we do then we will rename the presenter accordingly.