Create table body presenter (#476)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-1

Similar to how we generate transaction and customer files, we will be piping the contents of a db table row by row through a presenter which formats the row into sequentially-numbered columns, and then through a presenter which formats the row as CSV, and finally the row will be piped into a file.

We start developing this process by creating the table body presenter. Note that we may end up reusing this for the table head; if we do then we will rename the presenter accordingly.