Make `index` optional in `StreamTransformUsingPresenterService`
Currently, if no value for `indexStart` is passed to `StreamTransformUsingPresenterService` then it defaults to `0`. However, in the existing code for generating a transaction or customer file, we always pass a value and don't rely on the default of `0`. We can therefore make a change to the default behaviour; now, if no value is passed, the index will not be included in the output at all.