Create TransformTableToFile service (#477)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-1

We want to take a database table and pipe its contents to a CSV file. We have a presenter which will format the body of the table; the next stage is to create the service which will use this to transform a query into a CSV file.

This change introduces TransformTableToFileService, which adapts the existing TransformRecordsToFileService. We will later be looking at refactoring in order to share common functions; but for now we have this as an entirely separate service.

We make a small change to StreamTransformUsingPresenterService to allow us to reuse it here -- it does everything we need except it also adds in the index number needed for producing customer and transaction files. This has the effect of adding the index value to our output CSV (as the table file presenter simply passes on all the data it receives). We therefore modify StreamTransformUsingPresenterService so that if no starting index value is passed to it, the index is not added to the output.



* Create TransformTableToFile service

* First pass at service

This first pass will write the supplied column names and the output of the query to a file. However as we are reusing `StreamTransformUsingPresenterService` which also adds the line number to the stream, we find that each line ends with `"0"`.

* Make `index` optional in `StreamTransformUsingPresenterService`

Currently, if no value for `indexStart` is passed to `StreamTransformUsingPresenterService` then it defaults to `0`. However, in the existing code for generating a transaction or customer file, we always pass a value and don't rely on the default of `0`. We can therefore make a change to the default behaviour; now, if no value is passed, the index will not be included in the output at all.

* Rename presenter

While developing the service it became clear that `TableBodyPresenter` would be used for both the body and the header line. We therefore renamed it to `TableFilePresenter` to better reflect its purpose.