Create table export service (#480)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-1

We have TransformTableToFileService which can output a table to a file; we now need a service which accepts a table name and generates the required parameters (selecting query, column names, and output filename) which it then uses to call TransformTableToFileService.

A note on testing: it creates a customer change along with a customer file (so the customer change record can be populated with the file id) then exports the table and compares the customer change db record to the exported file. I would have preferred to completely mock the db calls to control exactly what data is fed to TransformTableToFileService but ran into difficulties with this. This therefore may be something to revisit in future.