Include licence no. for min. charge in trans file
https://eaflood.atlassian.net/browse/CMEA-12

The WRLS team on behalf of their users have told us that the licence no. needs to be included in the transaction file we generate. This is so that the customer can work out how the minimum charge was calculated for each individual licence on an invoice.

Prior to this change there was a requirement to not include the licence no. if the transaction was subject to minimum charge or a compensation charge. This change updates the logic to ensures it's only blank when the transaction is a compensation charge.