Reject requests with inverted quotation marks
https://eaflood.atlassian.net/browse/CMEA-11

During initial business acceptance testing of the WRLS system and its integration with the CM API found a transaction file that was rejected by SOP.

This was tracked down to the line description containing 'inverted quotation marks'. It wasn't possible to get an exact example. But based on the description (and Google!) we have interpreted this as

- **“** [Left Double Quotation Mark](https://unicode-table.com/en/201C/) `U+201C`
- **”** [Right Double Quotation Mark](https://unicode-table.com/en/201D/) `U+201D`

This change adds these to our `InvalidCharactersPlugin` which will mean any `POST` request that contains them anywhere in its content will be rejected.