Refactor helpers to accept objects not ids
We change the helpers so that instead of passing in the id of the parent, we pass in the parent object itself. So for example, when creating a licence on in invoice, we pass in `invoice` rather than `invoice.id`. This means that when we need information about the parent (eg. the bill run id) we can directly take it from the parent, instead of having a query inside the helper that retrieves the parent object and then gets the info.