Further update to comments and error message
It was further highlighted that we were also missing the em dash from our error messaging. So, this time we took a pause and ensured not only the error message but the code comments accurately reflected what we were checking for.

It also adds links to the specific characters in question so there can be no doubt what we are checking for.