Rework bill run, licence and invoice creation test helpers (#490)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-42

We currently create bill runs, licences and invoices in various ways throughout our tests. This PR aims to introduce some consistency to how we do this.

We have created NewBillRunHelper, NewInvoiceHelper and NewLicenceHelper, each of which will have a refactored method to add that entity. The basic pattern of each is that the first parameter will be the 'parent' object (ie. for a licence, the parent will be an invoice; and for an invoice, the parent will be the bill run) and the second will be an overrides object.

If no parent is specified then the helper will automatically create a parent, which in turn will create a parent if needed. So for example, calling NewInvoiceHelper.add() will work its way up to the top level to create a bill run, then create an invoice on it, and finally create a licence on that invoice.

For bill runs, we consider the authorised system and regime to be the parents; leaving either of these blank when calling add() will automatically create it.

The default properties of the created entity can be individually overridden using the overrides object, allowing us to only specify the things we need to be different from the default (which should make the intent of creating the entity clearer). Note that this is only the case for the helper that is being directly called; so for example, it won't be possible to create an invoice and have it create a bill run with generated status. In this case we would either create the bill run with the required status then create the invoice on it; or we would create the invoice and then patch the bill run status.

To begin with, this PR creates the new helpers, and refactors validate_bill_run_licence.service.test.js to use them (as a way of testing/demonstrating how the new helpers work). We will create additional PRs to change the rest of the existing tests a few at a time.


* Create NewBillRunHelper

This new helper refactors the `addBillRun` method to create a regime and authorised system if not specified.

* Rework ValidateBillRunLicenceService test to use NewBillRunHelper

* Create NewInvoiceHelper

* Rework ValidateBillRunLicenceService test to use NewInvoiceRunHelper

* Remove stray `.only`

* Small refactor of bill run and invoice helpers

* Create NewLicenceHelper

* Rework ValidateBillRunLicenceService test to use NewLicenceHelper

* Fix incorrect invoice creation in NewLicenceHelper

* Further refactor ValidateBillRunLicenceService test

* Refactor helpers to accept objects not ids

We change the helpers so that instead of passing in the id of the parent, we pass in the parent object itself. So for example, when creating a licence on in invoice, we pass in `invoice` rather than `invoice.id`. This means that when we need information about the parent (eg. the bill run id) we can directly take it from the parent, instead of having a query inside the helper that retrieves the parent object and then gets the info.

* Rename add methods to `add()`

Each helper has a uniquely-named add method, ie. `addBillRun()`, `addLicence()` and `addInvoice()`. Since the entity type is already in the helper name we felt this repetition was unneeded so we standardise the methods as simply `add()`.