Reject requests with inverted quotation marks (#492)
https://eaflood.atlassian.net/browse/CMEA-11

During initial business acceptance testing of the WRLS system and its integration with the CM API found a transaction file that was rejected by SOP.

This was tracked down to the line description containing 'inverted quotation marks'. It wasn't possible to get an exact example. But based on the description (and Google!) we have interpreted this as

- **“** [Left Double Quotation Mark](https://unicode-table.com/en/201C/) `U+201C`
- **”** [Right Double Quotation Mark](https://unicode-table.com/en/201D/) `U+201D`

This change adds these to our `InvalidCharactersPlugin` which will mean any `POST` request that contains them anywhere in its content will be rejected.

**Notes

We've not added tests for _every_ invalid character before because we are confident the logic is the same for each. But to confirm our change we did add a test just to prove it's covering the new invalid characters.

We also spotted the comments and error message didn't actually match what was being checked so we tidied those up at the same time.