Remove invalid min. charge model tests (#494)
Whilst working on another PR we spotted a couple of tests for the invoice model which are invalid/irrelevant. We think they are just the result of a copy and paste that never got cleaned up.

This change removes them and adds a test for 'false' condition to be consistent with others in the file.