Small change to `RegimeHelper`
We encountered an issue when refactoring controller tests where creating multiple bill runs via the helpers would fail. This was because `NewBillRunHelper`, when called with no regime (as is the case when it is called by creating an invoice, licence or transaction without a parent) would create a new regime for the bill run. This would then result in an error as we would have multiple regimes with the same slug `wrls`, and hitting an endpoint in the `/wrls/` path would only correspond to one of these.

We resolved this by adding an extra check to `RegimeHelper`. It first checks whether a regime already exists with the provided slug; if it does then it simply returns that regime, otherwise it creates a new one.