Remove $minimumChargeInvoice instance method
We're using this in the `DeleteLicenceService` to denote what to set the matching flag to on the invoice when a licence is deleted.

However, we've determined the error is because of our faulty logic in this method. For the purposes of determining whether an invoice needs to be looked at for minimum charge we do use the `subjectToMinimumChargeCount`.

But if all transactions are for more than £25, then we won't create any minimum charge adjustments and for the purposes of the transaction file we don't consider it a minimum charge invoice. This is what the flag refers to. Hence, the current logic of just checking the `subjectToMinimumChargeCount` is wrong. We need to be checking if any minimum charge adjustments remain.

Note - This is gonna break things but we expect that until we implement the fix!