Update and correct tests
We're going to need to query the DB to see if an invoice has minimum charge adjustment transactions. But we know these only get created when the bill run is generated. So, rather than hitting the DB unneccessarily we'll be adding some guard clauses based on bill run status and invoice subject to minimum charge count.

First step then is to update our tests to cater for this and ensure we are covering all scenarios (and all logic branches for code coverage).

When we did this we found the tests were still failing. We tracked this down to a misunderstanding about the `RulesServiceHelper`. We thought we were passing through £ values that would eventually get translated through to pennies. Turns out the helper is expecting pennies to be passed through which it then converts to pounds itself. So, prior to this change both licences were being treated as needing minimum charge adjustments. Doh!