Add exceptions to adding numbers
For the most part we want to add numbers when we update them -- counts and values, for example. However there are some things stored as numbers that we would want to replace, such as bill run number and financial year. We therefore amend our `update` methods to check for these exceptions.