Fix invoice min. charge flag when licence deleted (#493)
https://eaflood.atlassian.net/browse/CMEA-82

When an invoice has a mix of min charge licences there is a scenario where the `minimumChargeInvoice` flag on the `invoice` is not set correctly after a licence has been deleted

- create a bill run
- add a 'subject to minimum charge' transaction < £25 for licence `L001`, customer `C01`, FY `2014`
- add a 'subject to minimum charge' transaction for > £25 for licence `L002`, customer `C01`, FY `2014`
- generate the bill run
- **DELETE** `L001`

The Invoice level Min Charge flag is not updated to `false`.

This change fixes the issue.

* About the issue

We're using this in the `DeleteLicenceService` to denote what to set the matching flag to on the invoice when a licence is deleted.

However, we've determined the error is because of our faulty logic in the `$minimumChargeInvoice` invoice instance method. For the purposes of determining whether an invoice needs to be looked at for minimum charge we were using the `subjectToMinimumChargeCount`.

But if all transactions are for more than £25, then we won't create any minimum charge adjustments and for the purposes of the transaction file we don't consider it a minimum charge invoice. This is what the flag _actually_ refers to. Hence, the logic of just checking the `subjectToMinimumChargeCount` was wrong. We needed to be checking if any minimum charge adjustments remain.