Move flag-setting functionality of `NewInvoiceHelper` into `refreshFlags` method
Previously, the `add` method of `NewInvoiceHelper` set the invoice's zero value and deminimis flags based on its current counts and values. After some thought, we realised this wasn't ideal. Firstly, it means that if we want to test the behaviour of an invoice based on its flags, we would need to craft an invoice with the exact values required rather than simply setting the flags. We would find that if we did set the flags when creating an invoice, they would be overridden by the "proper" flags, which could be confusing. There was also uncertainty how this would work when calling the `update` method; should the flags be set at this point as well?

The conclusion we came to was that we should move the flag-setting functionality into its own method, `refreshFlags`; this way, we have complete control over whether or not we want to directly set the flags, and calling the `add` and `update` methods doesn't result in the flags being set when we're not expecting it.