Test helpers cascade changes upwards (#497)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-4

The new test helpers we have created will allow us to greatly streamline our tests. For example, if we want to test a transaction in isolation then we only need to create the transaction and everything else (bill run, invoice, licence, even authorised system and regime) will be handled for us.

However, if we do this then the bill run, invoice and licence will all be blank, which does not reflect the reality of what happens when we create things. We are therefore adding an update method to their helpers which will allow us to pass in an object of values to be added to the existing values. This is then used by the add method to cascade changes upwards, so if a transaction is created with certain values then these values will be added to the licence, invoice and bill run.

We also introduce test helper tests as part of this change. These are purely intended for us to test and demonstrate that the functionality works; as we start to update the existing tests to use the new helpers, we can ultimately delete the tests when we are satisfied that they are sufficiently tested through use.


* Add `update` method to `NewBillRunHelper`

* Add `update` method to `NewInvoiceHelper`

* Add `update` method to `NewLicenceHelper`

* `NewInvoiceHelper` updates bill run when adding

Once an invoice has been created in the `add` method of `NewInvoiceHelper`, we extract the relevant values and pass them to the `update` method of `NewBillRunHelper` in order to update the bill run before we return the new invoice.

* Refactor `NewInvoiceHelper` tests

* `NewLicenceHelper` updates invoice when adding

* `NewTransactionHelper` updates licence when adding

We use the existing `CreateTransactionTallyService` to calculate the fields to be updated at the licence level as this gives the exact details we need in the `insertData` of the returned object, and is thoroughly tested for all scenarios.

* Remove stray .only

* Change `update` methods to overwrite non-numbers

The `update` methods of the new helpers will now handle non-numbers (ie. strings, booleans etc.) -- any numbers passed in will be added to the existing value as before, while any non-numbers will replace the existing value entirely

* Add exceptions to adding numbers

For the most part we want to add numbers when we update them -- counts and values, for example. However there are some things stored as numbers that we would want to replace, such as bill run number and financial year. We therefore amend our `update` methods to check for these exceptions.

* Move flag-setting functionality of `NewInvoiceHelper` into `refreshFlags` method

Previously, the `add` method of `NewInvoiceHelper` set the invoice's zero value and deminimis flags based on its current counts and values. After some thought, we realised this wasn't ideal. Firstly, it means that if we want to test the behaviour of an invoice based on its flags, we would need to craft an invoice with the exact values required rather than simply setting the flags. We would find that if we did set the flags when creating an invoice, they would be overridden by the "proper" flags, which could be confusing. There was also uncertainty how this would work when calling the `update` method; should the flags be set at this point as well?

The conclusion we came to was that we should move the flag-setting functionality into its own method, `refreshFlags`; this way, we have complete control over whether or not we want to directly set the flags, and calling the `add` and `update` methods doesn't result in the flags being set when we're not expecting it.

* Rename `add` methods to `create`

We decided that `create` better reflects the intent of the helper `add` methods.